refactor: add LunchedEffects in another file
